Switch to double precision, implement intersections on CPU

+++ scene.h
@@ -13,13 +13,13 @@ struct Object
 struct Scene
 {
     int px_width, px_height;
-    float vertical_fov;
+    double vertical_fov;
     glm::vec3 bg_color;
 
     std::vector<Object> objects;
 
-    Scene(int px_width, int px_height, float vertical_fov_deg, const glm::vec3& bg_color = glm::vec3(0.0f));
-    float aspect_ratio() const { return (float) px_width / (float) px_height; }
+    Scene(int px_width, int px_height, double vertical_fov_deg, const glm::vec3& bg_color = glm::vec3(0.0f));
+    double aspect_ratio() const { return (double) px_width / px_height; }
 };
 
 #endif //CUDA_RAY_TRACER_SCENE_H
